+++ examples/createCustomer.js
@@ -1,6 +1,6 @@
 const tuneuptechnology = require('../src/tuneuptechnology')
 
-const client = new tuneuptechnology.Client(process.env.API_EMAIL, process.env.API_KEY)
+const client = new tuneuptechnology.client(process.env.API_EMAIL, process.env.API_KEY)
 
 const data = {
     firstname: 'Jake',