+++ examples/deleteCustomer.js
@@ -1,5 +1,5 @@
 const tuneuptechnology = require('tuneuptechnology')
 
-const client = new tuneuptechnology.Client(process.env.API_EMAIL, process.env.API_KEY)
+const client = new tuneuptechnology.client(process.env.API_EMAIL, process.env.API_KEY)
 
-client.Customers.delete(23).then(console.log).catch(console.log)
+client.customers.delete(23).then(console.log).catch(console.log)