+++ examples/retrieveCustomers.js
@@ -1,5 +1,5 @@
 const tuneuptechnology = require('tuneuptechnology')
 
-const client = new tuneuptechnology.Client(process.env.API_EMAIL, process.env.API_KEY)
+const client = new tuneuptechnology.client(process.env.API_EMAIL, process.env.API_KEY)
 
-client.Customers.all().then(console.log).catch(console.log)
+client.customers.all().then(console.log).catch(console.log)