+++ test/client.test.js
@@ -4,6 +4,6 @@ const tuneuptechnology = require('../index')
 
 describe('Client', function () {
     it('throws an error when missing credentials', async function () {
-        assert.throws(() => { new tuneuptechnology.Client(null, null, 'http://tuneapp.localhost/api') }, Error)
+        assert.throws(() => { new tuneuptechnology.client(null, null, 'http://tuneapp.localhost/api') }, Error)
     })
 })