+++ test/customers.test.js
@@ -10,7 +10,7 @@ describe('Customers', function () {
         const { server } = this.polly
         setupPolly.stripCreds(server)
 
-        const client = new tuneuptechnology.Client(process.env.API_EMAIL, process.env.API_KEY, 'http://tuneapp.localhost/api')
+        const client = new tuneuptechnology.client(process.env.API_EMAIL, process.env.API_KEY, 'http://tuneapp.localhost/api')
 
         const data = {
             firstname: 'Jake',
@@ -31,7 +31,7 @@ describe('Customers', function () {
         const { server } = this.polly
         setupPolly.stripCreds(server)
 
-        const client = new tuneuptechnology.Client(process.env.API_EMAIL, process.env.API_KEY, 'http://tuneapp.localhost/api')
+        const client = new tuneuptechnology.client(process.env.API_EMAIL, process.env.API_KEY, 'http://tuneapp.localhost/api')
 
         const response = await client.customers.retrieve(1)
 
@@ -42,7 +42,7 @@ describe('Customers', function () {
         const { server } = this.polly
         setupPolly.stripCreds(server)
 
-        const client = new tuneuptechnology.Client(process.env.API_EMAIL, process.env.API_KEY, 'http://tuneapp.localhost/api')
+        const client = new tuneuptechnology.client(process.env.API_EMAIL, process.env.API_KEY, 'http://tuneapp.localhost/api')
 
         const response = await client.customers.all()
 
@@ -53,7 +53,7 @@ describe('Customers', function () {
         const { server } = this.polly
         setupPolly.stripCreds(server)
 
-        const client = new tuneuptechnology.Client(process.env.API_EMAIL, process.env.API_KEY, 'http://tuneapp.localhost/api')
+        const client = new tuneuptechnology.client(process.env.API_EMAIL, process.env.API_KEY, 'http://tuneapp.localhost/api')
 
         const data = {
             firstname: 'Jake',
@@ -74,7 +74,7 @@ describe('Customers', function () {
         const { server } = this.polly
         setupPolly.stripCreds(server)
 
-        const client = new tuneuptechnology.Client(process.env.API_EMAIL, process.env.API_KEY, 'http://tuneapp.localhost/api')
+        const client = new tuneuptechnology.client(process.env.API_EMAIL, process.env.API_KEY, 'http://tuneapp.localhost/api')
 
         const response = await client.customers.delete(1)